Increase maxtick type from 8bit to 64bit.

+++ citrine.h
@@ -532,8 +532,8 @@ ctr_object* ctr_command_forbid_include(ctr_object* myself, ctr_argument* ctr_arg
 ctr_object* ctr_command_countdown(ctr_object* myself, ctr_argument* ctr_argumentList);
 void ctr_check_permission( uint8_t operationID );
 uint8_t ctr_command_security_profile;
-uint8_t ctr_command_tick;
-uint8_t ctr_command_maxtick;
+uint64_t ctr_command_tick;
+uint64_t ctr_command_maxtick;
 #define CTR_SECPRO_NO_SHELL 1
 #define CTR_SECPRO_NO_FILE_WRITE 2
 #define CTR_SECPRO_NO_FILE_READ 4
