+++ system.c
@@ -557,7 +557,7 @@ ctr_object* ctr_command_forbid_include( ctr_object* myself, ctr_argument* argume
  */
 ctr_object* ctr_command_countdown( ctr_object* myself, ctr_argument* argumentList ) {
 	ctr_command_security_profile |= CTR_SECPRO_COUNTDOWN;
-	ctr_command_maxtick = (uint8_t) ctr_internal_cast2number( argumentList->object )->value.nvalue;
+	ctr_command_maxtick = (uint64_t) ctr_internal_cast2number( argumentList->object )->value.nvalue;
 	return myself;
 }
